+++ CHANGELOG.md
@@ -7,6 +7,10 @@ This file keeps track of all changes to this project. This project follows [sema
 
 ## [UNRELEASED]
 
+### Changed
+- The `performance-analysis` module has been extracted to its own project: https://github.com/FaKeller/performance.
+
+
 ## v0.5.1 - 2016-10-04 [Maintenance Release]
 
 ### Changed